+++ app/Base/layouts/ProtectedLayout/index.tsx
@@ -1,7 +1,6 @@
 import {
     ActionIcon,
     AppShell,
-    Center,
     Divider,
     Flex,
     Header,
